python-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mlbb-automation"
version = "0.1.0"
description = "Android device farm automation for Mobile Legends: Bang Bang"
requires-python = ">=3.10"
dependencies = [
    "Appium-Python-Client>=3.1.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "Pillow>=10.1.0",
    "click>=8.1.7",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
    "urllib3>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
mlbb-automation = "mlbb_automation.__main__:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["mlbb_automation*"]
